+++ compiler-core/src/format/tests/imports.rs
@@ -7,3 +7,19 @@ fn types_and_values() {
 "
     );
 }
+
+#[test]
+fn discarded_import() {
+    assert_format!(
+        "import one/two as _three
+"
+    );
+}
+
+#[test]
+fn discarded_import_with_unqualified() {
+    assert_format!(
+        "import one/two.{type Abc, Bcd, abc} as _three
+"
+    );
+}