+++ compiler-core/src/javascript/tests.rs
@@ -33,10 +33,19 @@ mod use_;
 pub static CURRENT_PACKAGE: &str = "thepackage";
 
 #[macro_export]
+macro_rules! assert_js_with_multiple_imports {
+    ($(($name:literal, $module_src:literal)),+; $src:literal) => {
+        let output =
+            $crate::javascript::tests::compile_js($src, vec![$((CURRENT_PACKAGE, $name, $module_src)),*]);
+        insta::assert_snapshot!(insta::internals::AutoName, output, $src);
+    };
+}
+
+#[macro_export]
 macro_rules! assert_js {
     (($dep_package:expr, $dep_name:expr, $dep_src:expr), $src:expr $(,)?) => {{
         let output =
-            $crate::javascript::tests::compile_js($src, Some(($dep_package, $dep_name, $dep_src)));
+            $crate::javascript::tests::compile_js($src, vec![($dep_package, $dep_name, $dep_src)]);
         insta::assert_snapshot!(insta::internals::AutoName, output, $src);
     }};
 
@@ -47,12 +56,12 @@ macro_rules! assert_js {
     }};
 
     ($src:expr $(,)?) => {{
-        let output = $crate::javascript::tests::compile_js($src, None);
+        let output = $crate::javascript::tests::compile_js($src, vec![]);
         insta::assert_snapshot!(insta::internals::AutoName, output, $src);
     }};
 
     ($src:expr, $js:expr $(,)?) => {{
-        let output = $crate::javascript::tests::compile_js($src, None);
+        let output = $crate::javascript::tests::compile_js($src, vec![]);
         assert_eq!(($src, output), ($src, $js.to_string()));
     }};
 }
@@ -61,17 +70,17 @@ macro_rules! assert_js {
 macro_rules! assert_ts_def {
     (($dep_package:expr, $dep_name:expr, $dep_src:expr), $src:expr $(,)?) => {{
         let output =
-            $crate::javascript::tests::compile_ts($src, Some(($dep_package, $dep_name, $dep_src)));
+            $crate::javascript::tests::compile_ts($src, vec![($dep_package, $dep_name, $dep_src)]);
         insta::assert_snapshot!(insta::internals::AutoName, output, $src);
     }};
 
     ($src:expr $(,)?) => {{
-        let output = $crate::javascript::tests::compile_ts($src, None);
+        let output = $crate::javascript::tests::compile_ts($src, vec![]);
         insta::assert_snapshot!(insta::internals::AutoName, output, $src);
     }};
 }
 
-pub fn compile(src: &str, dep: Option<(&str, &str, &str)>) -> TypedModule {
+pub fn compile(src: &str, deps: Vec<(&str, &str, &str)>) -> TypedModule {
     let mut modules = im::HashMap::new();
     let ids = UniqueIdGenerator::new();
     // DUPE: preludeinsertion
@@ -84,24 +93,24 @@ pub fn compile(src: &str, dep: Option<(&str, &str, &str)>) -> TypedModule {
     );
     let mut direct_dependencies = std::collections::HashMap::from_iter(vec![]);
 
-    if let Some((dep_package, dep_name, dep_src)) = dep {
+    deps.iter().for_each(|(dep_package, dep_name, dep_src)| {
         let parsed = crate::parse::parse_module(dep_src).expect("dep syntax error");
         let mut ast = parsed.module;
-        ast.name = dep_name.into();
+        ast.name = (*dep_name).into();
         let dep = crate::analyse::infer_module::<()>(
             Target::JavaScript,
             &ids,
             ast,
             Origin::Src,
-            &dep_package.into(),
+            &((*dep_package).into()),
             &modules,
             &TypeWarningEmitter::null(),
             &std::collections::HashMap::new(),
         )
         .expect("should successfully infer");
-        let _ = modules.insert(dep_name.into(), dep.type_info);
-        let _ = direct_dependencies.insert(dep_package.into(), ());
-    }
+        let _ = modules.insert((*dep_name).into(), dep.type_info);
+        let _ = direct_dependencies.insert((*dep_package).into(), ());
+    });
 
     let parsed = crate::parse::parse_module(src).expect("syntax error");
     let mut ast = parsed.module;
@@ -119,13 +128,13 @@ pub fn compile(src: &str, dep: Option<(&str, &str, &str)>) -> TypedModule {
     .expect("should successfully infer")
 }
 
-pub fn compile_js(src: &str, dep: Option<(&str, &str, &str)>) -> String {
-    let ast = compile(src, dep);
+pub fn compile_js(src: &str, deps: Vec<(&str, &str, &str)>) -> String {
+    let ast = compile(src, deps);
     let line_numbers = LineNumbers::new(src);
     module(&ast, &line_numbers, Utf8Path::new(""), &"".into()).unwrap()
 }
 
-pub fn compile_ts(src: &str, dep: Option<(&str, &str, &str)>) -> String {
-    let ast = compile(src, dep);
+pub fn compile_ts(src: &str, deps: Vec<(&str, &str, &str)>) -> String {
+    let ast = compile(src, deps);
     ts_declaration(&ast, Utf8Path::new(""), &src.into()).unwrap()
 }