+++ compiler-core/src/parse.rs
@@ -58,12 +58,11 @@ use crate::analyse::Inferred;
 use crate::ast::{
     Arg, ArgNames, AssignName, Assignment, AssignmentKind, BinOp, BitArrayOption, BitArraySegment,
     CallArg, Clause, ClauseGuard, Constant, CustomType, Definition, Function, HasLocation, Import,
-    ImportName, Module, ModuleConstant, Pattern, RecordConstructor, RecordConstructorArg,
-    RecordUpdateSpread, SrcSpan, Statement, TargetedDefinition, TodoKind, TypeAlias, TypeAst,
-    TypeAstConstructor, TypeAstFn, TypeAstHole, TypeAstTuple, TypeAstVar, UnqualifiedImport,
-    UntypedArg, UntypedClause, UntypedClauseGuard, UntypedConstant, UntypedDefinition, UntypedExpr,
-    UntypedModule, UntypedPattern, UntypedRecordUpdateArg, UntypedStatement, Use, UseAssignment,
-    CAPTURE_VARIABLE,
+    Module, ModuleConstant, Pattern, RecordConstructor, RecordConstructorArg, RecordUpdateSpread,
+    SrcSpan, Statement, TargetedDefinition, TodoKind, TypeAlias, TypeAst, TypeAstConstructor,
+    TypeAstFn, TypeAstHole, TypeAstTuple, TypeAstVar, UnqualifiedImport, UntypedArg, UntypedClause,
+    UntypedClauseGuard, UntypedConstant, UntypedDefinition, UntypedExpr, UntypedModule,
+    UntypedPattern, UntypedRecordUpdateArg, UntypedStatement, Use, UseAssignment, CAPTURE_VARIABLE,
 };
 use crate::build::Target;
 use crate::parse::extra::ModuleExtra;
@@ -1986,13 +1985,16 @@ where
         // Parse as_name
         let mut as_name = None;
         if let Some((as_start, _)) = self.maybe_one(&Token::As) {
-            let (_, name, e) = self.expect_name()?;
-            let location = SrcSpan {
-                start: as_start,
-                end: e,
-            };
-            as_name = Some(ImportName { location, name });
+            let (_, name, e) = self.expect_assign_name()?;
+
             end = e;
+            as_name = Some((
+                name,
+                SrcSpan {
+                    start: as_start,
+                    end,
+                },
+            ));
         }
 
         Ok(Some(Definition::Import(Import {