+++ compiler-core/src/type_/expression.rs
@@ -1416,6 +1416,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
             // Register this imported module as having been used, to inform
             // warnings of unused imports later
             let _ = self.environment.unused_modules.remove(module_alias);
+            let _ = self.environment.unused_module_aliases.remove(module_alias);
 
             (module.name.clone(), constructor.clone())
         };
@@ -1781,6 +1782,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 // Register the module as having been used if it was imported
                 if let Some(ref module) = &module {
                     _ = self.environment.unused_modules.remove(module);
+                    _ = self.environment.unused_module_aliases.remove(module);
                 }
 
                 // Type check the record constructor
@@ -1825,6 +1827,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 // Register the module as having been used if it was imported
                 if let Some(ref module) = &module {
                     _ = self.environment.unused_modules.remove(module);
+                    _ = self.environment.unused_module_aliases.remove(module);
                 }
 
                 let constructor = self.infer_value_constructor(&module, &name, &location)?;
@@ -1948,6 +1951,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 // Register the module as having been used if it was imported
                 if let Some(ref module) = &module {
                     _ = self.environment.unused_modules.remove(module);
+                    _ = self.environment.unused_module_aliases.remove(module);
                 }
 
                 // Infer the type of this constant